// src/presentation/Enums/ArmazonesEnum.ts
export enum ArmazonesEnum {
  codigo = 1,
  ubicacion = 2,
  stock_minimo = 3,
  cantidad_exhibida = 4,
  cantidad_reservada = 5,
  stock_disponible = 6,
  stock_total = 7,
  armazon_tipo_id = 8,
  armazon_tipo = 9,
  marca_id = 10,
  marca = 11,
  modelo = 12,
  color = 13,
  armazon_material_id = 14,
  armazon_material = 15,
  aro = 16,
  puente = 17,
  diagonal = 18,
  brazo = 19,
  armazon_uso_id = 20,
  armazon_uso = 21,
  dp_minima = 22,
  dp_maxima = 23,
  codigo_fab = 24,
}